save new Instrument and some bug fixed

+++ src/main/java/com/gitlab/amirmehdi/service/dto/OptionStockWatch.java
@@ -22,7 +22,7 @@ public class OptionStockWatch {
     @JsonProperty("BSPrice")
     private int bsPrice;
     @JsonProperty("TradeVolume")
-    private int tradeVolume;
+    private long tradeVolume;
     @JsonProperty("TradeCount")
     private int tradeCount;
     @JsonProperty("TradeValue")
